+++ CHANGELOG.md
@@ -78,6 +78,7 @@ high state of flux, you're at risk of it changing without notice.
     - deprecate `getJoinMonoid`, use `max` instead (@gcanti)
     - deprecate `getDualMonoid`, use `reverse` instead (@gcanti)
     - deprecate `getStructMonoid`, use `struct` instead (@gcanti)
+    - deprecate `getTupleMonoid`, use `tuple` instead (@gcanti)
   - `NonEmptyArray`
     - deprecate `fold`, use `concatAll` instead (@gcanti)
   - `Option`
@@ -265,6 +266,7 @@ high state of flux, you're at risk of it changing without notice.
     - add `max` (@gcanti)
     - add `reverse` (@gcanti)
     - add `struct` (@gcanti)
+    - add `tuple` (@gcanti)
   - `NonEmptyArray`
     - add `Pointed` instance (@gcanti)
     - add `Apply` instance (@gcanti)